Broke down entity getters
